Setting up (#1)
Initial set up of all files. This is rather rough-and-ready, and could use some polish. However, it works with MacOS and Raspberry Pi hosts.

I have had to use `pyserial` to handle the communications on the host; when I tried `io.TextIOWrapper()` around the entry in `/dev/` directly, something strange was happening and the Pico was getting confused (I suspect dropping into the REPL, but I couldn't confirm this).